+++ src/components/Molecules/VideoCarousel/VideoCarousel.module.scss
@@ -1,0 +1,20 @@
+.card {
+  height: 300px;
+  width: 250px;
+  margin: 0 auto;
+  transition: all .2s ease-in-out;
+
+  &:hover {
+    transform: translateY(-10px);
+  }
+
+  .cardMedia {
+    width: 100%;
+    height: 120px;
+  }
+
+  .publishedAt {
+    color: #bbb;
+    margin-top: 5px;
+  }
+}
