+++ src/components/Molecules/VideoDialog/VideoDialog.module.scss
@@ -1,0 +1,13 @@
+.videoFrame {
+  border: 0;
+  width: 100%;
+}
+
+.dialogDescription {
+  padding-top: 2rem;
+}
+
+.publishedText {
+  padding-top: 2rem;
+  color: #bbb;
+}
