+++ src/components/Molecules/VideosSection/VideosSection.tsx
@@ -1,36 +1,13 @@
-import React, { SFC } from 'react';
+import React from 'react';
 import compose from 'recompose/compose';
 import Section from '../../../Section/Section';
 import { translate, TranslationFunction } from 'react-i18next';
 import styles from './VideosSection.module.scss';
 import { VideoItem } from '../../../containers/EventsContainer/fetchYouTubePlaylist';
-import {
-  Typography,
-  Card,
-  CardMedia,
-  CardContent,
-  Dialog,
-  DialogContent,
-  DialogActions,
-  Button,
-} from '@material-ui/core';
-import Carousel from 'nuka-carousel';
-import {
-  leftControls,
-  rightControls,
-  bottomControls,
-} from '../TeamCarousel/renderControls';
-import i18n from '../../../i18n';
+import { Typography, Button } from '@material-ui/core';
 import { YOUTUBE_CODESTAR_CHANNEL_ID } from '../../../constants';
-
-const formatDate = (date: Date, language: string) => {
-  const locale = language === 'nl' ? 'nl-NL' : 'en-US';
-  return new Date(date).toLocaleDateString(locale, {
-    year: 'numeric',
-    month: 'long',
-    day: 'numeric',
-  });
-};
+import { VideoCarousel } from '../VideoCarousel/VideoCarousel';
+import { VideoDialog } from '../VideoDialog/VideoDialog';
 
 interface IPropsInner {
   t: TranslationFunction;
@@ -120,119 +97,6 @@ export class VideosSection extends React.Component<
     });
   };
 }
-
-type VideoCarouselProps = Readonly<{
-  videos: VideoItem[];
-  cardWidth: number;
-  onSelectVideo: (video: VideoItem) => void;
-}>;
-
-const VideoCarousel: SFC<VideoCarouselProps> = props => {
-  return (
-    <Carousel
-      slideWidth={`${props.cardWidth}px`}
-      wrapAround
-      cellAlign="left"
-      slidesToScroll="auto"
-      renderCenterLeftControls={leftControls}
-      renderCenterRightControls={rightControls}
-      renderBottomCenterControls={bottomControls}
-    >
-      {props.videos.map(video => (
-        <Card
-          key={video.id}
-          className={`${styles.card} my-3`}
-          onClick={() => props.onSelectVideo(video)}
-        >
-          <CardMedia
-            className={styles.cardMedia}
-            image={video.thumbnails.medium.url}
-            title={video.title}
-          />
-          <CardContent>
-            <Typography variant="headline" component="h3" title={video.title}>
-              <Truncate text={video.title} limit={40} />
-            </Typography>
-            <p className={styles.publishedAt}>
-              {formatDate(video.publishedAt, i18n.language)}
-            </p>
-          </CardContent>
-        </Card>
-      ))}
-    </Carousel>
-  );
-};
-
-type VideoDialogProps = Readonly<{
-  video: VideoItem | false;
-  fullScreen: boolean;
-  onClose: () => void;
-  t: TranslationFunction;
-}>;
-
-const VideoDialog: SFC<VideoDialogProps> = ({
-  video,
-  fullScreen,
-  onClose,
-  t,
-}) => {
-  const content =
-    video !== false ? (
-      <>
-        <DialogContent style={{ overflow: 'visible' }}>
-          <h1>{video.title}</h1>
-        </DialogContent>
-        <iframe
-          className={styles.videoFrame}
-          width="560"
-          height="315"
-          src={`https://www.youtube.com/embed/${video.id}`}
-          allow="accelerometer; autoplay; encrypted-media; gyroscope; picture-in-picture"
-        />
-        <DialogContent>
-          {video.description.length ? (
-            <div className={styles.dialogDescription}>
-              {video.description.map(paragraph => (
-                <p>{paragraph}</p>
-              ))}
-            </div>
-          ) : null}
-          <div className={styles.publishedText}>
-            {t('PUBLISHED_ON')} {formatDate(video.publishedAt, i18n.language)}
-          </div>
-        </DialogContent>
-        <DialogActions>
-          <Button variant="contained" onClick={onClose} color="primary">
-            {t('EVENTS_CLOSE_BUTTON')}
-          </Button>
-        </DialogActions>
-      </>
-    ) : (
-      <>-No video selected-</>
-    );
-
-  return (
-    <Dialog
-      fullScreen={fullScreen}
-      open={video !== false}
-      onClose={onClose}
-      scroll={fullScreen ? 'paper' : 'body'}
-    >
-      {content}
-    </Dialog>
-  );
-};
-
-type TruncateProps = Readonly<{
-  limit: number;
-  text: string;
-}>;
-
-const Truncate: SFC<TruncateProps> = ({ limit, text }) => {
-  const textToRender =
-    text.length > limit ? `${text.substr(0, limit)}...` : text;
-  return <>{textToRender}</>;
-};
 
 export default compose<VideosSectionProps, IPropsOuter>(
   translate(['events'], { wait: true })
